Enhancing mark to allow transitionId definition.
In this change, i've enhanced the mark command to allow a transitionId to be passed in from the command line. This allows for greater scripting / automation possibilities.

We do not get the available transitions or ask the user to confirm their change for performance reasons but we assume a 502 error returned is a failed transition.

To enable this change I had to add error handling to transitions.doTransition which mean't changing the signature of the functions transitions. start, stop, review, done and invalid to accept a cb parameter and use accordingly inline with other areas of the cli.